hopefully a serverside crash fix
