+++ src/main/java/com/hbm/handler/imc/IMCHandlerNHNEI.java
@@ -11,64 +11,10 @@ import net.minecraft.nbt.NBTTagCompound;
 
 import java.util.ArrayList;
 
+import static com.hbm.main.ClientProxy.handlerList;
+
 public class IMCHandlerNHNEI {
-    
-    public static ArrayList<TemplateRecipeHandler> handlerList() {
-        ArrayList<TemplateRecipeHandler> handlers = new ArrayList<>();
-        
-        handlers.add(new AlloyFurnaceRecipeHandler());
-        handlers.add(new ShredderRecipeHandler());
-        handlers.add(new PressRecipeHandler());
-        handlers.add(new CentrifugeRecipeHandler());
-        handlers.add(new GasCentrifugeRecipeHandler());
-        handlers.add(new BreederRecipeHandler());
-        handlers.add(new CyclotronRecipeHandler());
-        handlers.add(new AssemblerRecipeHandler());
-        handlers.add(new RefineryRecipeHandler());
-        handlers.add(new VacuumRecipeHandler());
-        handlers.add(new CrackingHandler());
-        handlers.add(new ReformingHandler());
-        handlers.add(new HydrotreatingHandler());
-        handlers.add(new BoilerRecipeHandler());
-        handlers.add(new ChemplantRecipeHandler());
-        handlers.add(new CrystallizerRecipeHandler());
-        handlers.add(new BookRecipeHandler());
-        handlers.add(new FusionRecipeHandler());
-        handlers.add(new HadronRecipeHandler());
-        handlers.add(new SILEXRecipeHandler());
-        handlers.add(new SmithingRecipeHandler());
-        handlers.add(new AnvilRecipeHandler());
-        handlers.add(new FuelPoolHandler());
-        handlers.add(new FluidRecipeHandler());
-        handlers.add(new RadiolysisRecipeHandler());
-        handlers.add(new CrucibleSmeltingHandler());
-        handlers.add(new CrucibleAlloyingHandler());
-        handlers.add(new CrucibleCastingHandler());
-        handlers.add(new ToolingHandler());
-        handlers.add(new ConstructionHandler());
 
-        //universal boyes
-        handlers.add(new ZirnoxRecipeHandler());
-        if(VersatileConfig.rtgDecay()) {
-            handlers.add(new RTGRecipeHandler());
-        }
-        handlers.add(new LiquefactionHandler());
-        handlers.add(new SolidificationHandler());
-        handlers.add(new CokingHandler());
-        handlers.add(new FractioningHandler());
-        handlers.add(new BoilingHandler());
-        handlers.add(new CombinationHandler());
-        handlers.add(new SawmillHandler());
-        handlers.add(new MixerHandler());
-        handlers.add(new OutgasserHandler());
-        handlers.add(new ElectrolyserFluidHandler());
-        handlers.add(new ElectrolyserMetalHandler());
-        handlers.add(new AshpitHandler());
-        handlers.add(new ArcWelderHandler());
-        handlers.add(new ExposureChamberHandler());
-
-        return handlers;
-    }
     public static void IMCSender() {
 
         for (TemplateRecipeHandler handler: handlerList()) {
