+++ src/main/java/com/hbm/main/NEIConfig.java
@@ -25,64 +25,14 @@ import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.world.World;
 
+import static com.hbm.main.ClientProxy.handlerList;
+
 public class NEIConfig implements IConfigureNEI {
 	@Override
 	public void loadConfig() {
-		registerHandler(new AlloyFurnaceRecipeHandler());
-		registerHandler(new ShredderRecipeHandler());
-		registerHandler(new PressRecipeHandler());
-		registerHandler(new CentrifugeRecipeHandler());
-		registerHandler(new GasCentrifugeRecipeHandler());
-		registerHandler(new BreederRecipeHandler());
-		registerHandler(new CyclotronRecipeHandler());
-		registerHandler(new AssemblerRecipeHandler());
-		registerHandler(new RefineryRecipeHandler());
-		registerHandler(new VacuumRecipeHandler());
-		registerHandler(new CrackingHandler());
-		registerHandler(new ReformingHandler());
-		registerHandler(new HydrotreatingHandler());
-		registerHandler(new BoilerRecipeHandler());
-		registerHandler(new ChemplantRecipeHandler());
-		registerHandler(new CrystallizerRecipeHandler());
-		registerHandler(new BookRecipeHandler());
-		registerHandler(new FusionRecipeHandler());
-		registerHandler(new HadronRecipeHandler());
-		registerHandler(new SILEXRecipeHandler());
-		registerHandler(new SmithingRecipeHandler());
-		registerHandler(new AnvilRecipeHandler());
-		registerHandler(new FuelPoolHandler());
-		registerHandler(new RadiolysisRecipeHandler());
-		registerHandler(new CrucibleSmeltingHandler());
-		registerHandler(new CrucibleAlloyingHandler());
-		registerHandler(new CrucibleCastingHandler());
-		registerHandler(new ToolingHandler());
-		registerHandler(new ConstructionHandler());
-		
-		//universal boyes
-		registerHandler(new ZirnoxRecipeHandler());
-		if(VersatileConfig.rtgDecay()) {
-			registerHandler(new RTGRecipeHandler());
+		for (TemplateRecipeHandler handler: handlerList()) {
+			registerHandler(handler);
 		}
-		registerHandler(new LiquefactionHandler());
-		registerHandler(new SolidificationHandler());
-		registerHandler(new CokingHandler());
-		registerHandler(new FractioningHandler());
-		registerHandler(new BoilingHandler());
-		registerHandler(new CombinationHandler());
-		registerHandler(new SawmillHandler());
-		registerHandler(new MixerHandler());
-		registerHandler(new OutgasserHandler());
-		registerHandler(new ElectrolyserFluidHandler());
-		registerHandler(new ElectrolyserMetalHandler());
-		registerHandler(new AshpitHandler());
-		registerHandler(new ArcWelderHandler());
-		registerHandler(new ExposureChamberHandler());
-
-		for(MachineConfiguration conf : CustomMachineConfigJSON.niceList) registerHandlerBypass(new CustomMachineHandler(conf));
-		
-		//fluids
-		registerHandler(new FluidRecipeHandler());
-
 		//Some things are even beyond my control...or are they?
 		API.hideItem(ItemBattery.getEmptyBattery(ModItems.memory));
 		API.hideItem(ItemBattery.getFullBattery(ModItems.memory));
