+++ src/main/java/com/hbm/main/ServerProxy.java
@@ -34,6 +34,7 @@ public class ServerProxy {
 	public void registerItemRenderer() { }
 	public void registerEntityRenderer() { }
 	public void registerBlockRenderer() { }
+	public void handleNHNEICompat() { }
 	
 	public void particleControl(double x, double y, double z, int type) { }
